% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texts.R
\name{add_text}
\alias{add_text}
\title{Add text to database}
\usage{
add_text(lexadb, title = NULL)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}})}

\item{title}{The text title as a string.}
}
\value{
Nothing. Used for its side effects
}
\description{
This function creates a new text in the database, i.e. a new empty text
skeleton is written to disk, in the \verb{texts/} directory, for the user
to edit at will.
}
