% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{create_entry}
\alias{create_entry}
\title{Title}
\usage{
create_entry(
  lexadb = NULL,
  entry = NULL,
  gloss = NULL,
  part_of_speech = NULL,
  phon = NULL,
  morph_category = NULL,
  morph_type = NULL,
  definition = gloss,
  etymology = NULL,
  notes = NULL
)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{entry}{The entry as a string.}

\item{gloss}{The gloss as a string.}

\item{part_of_speech}{The part of speech as a string.}

\item{phon}{The phonetic transcription as a string.}

\item{morph_category}{The morphosyntactic category as a string (\code{"lexical"} or \code{"grammatical"}).}

\item{morph_type}{The type of morpheme as a string.}

\item{definition}{The definition of the entry as a string.}

\item{etymology}{The etymology of the entry as a string.}

\item{notes}{Further notes as a string.}
}
\value{
A list with entry id (\code{id}) and output string (\code{out}).
}
\description{
Title
}
