% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{import_lexicon}
\alias{import_lexicon}
\title{Import lexicon to existing database}
\usage{
import_lexicon(lexadb, path)
}
\arguments{
\item{lexadb}{A \code{lexadb} object as returned by \code{load_lexadb()}.}

\item{path}{The path to the lexicon .csv file as a string.}
}
\value{
A \code{lexadb} object.
}
\description{
It imports entries from a lexicon .csv file with specific column names.
See Details for file specifics.
}
\details{
The file should have the following columns (they can be omitted, but
if present they must be so named):
\itemize{
\item \code{entry}: the lexical entry, as it should appear in the head entry.
\item \code{phon}: phonetic transcription of the entry.
\item \code{morph_category}: category of entry (e.g. lexical vs grammatical).
\item \code{morph_type}: type of morpheme (e.g. root vs affix).
\item \code{part_of_speech}: part of speech of entry.
\item \code{class}: lexical class of entry (e.g. verbal conjugations, noun classes).
\item \code{gloss}: .
\item \code{etymology}:
\item \code{notes}:
}

Note that this list is temporary and it will change in the future.
}
