% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_lexicon_csv}
\alias{import_lexicon_csv}
\title{Import lexicon from a csv file}
\usage{
import_lexicon_csv(path, lexadb_path)
}
\arguments{
\item{path}{The path to the lexicon \code{.csv} file as a string.}

\item{lexadb_path}{The path to the Lexa database folder including the \verb{_lexadb} suffix.}
}
\value{
A new Lexa DB is created and the corresponding \code{lexadb} object is returned.
}
\description{
It imports entries from a \code{.csv} file with lexical data. The file must have
specific columns, see Details for file specifics. The lexicon is imported
into an existing Lexa database.
}
\details{
The file must have at least the following columns:
\itemize{
\item \code{lexeme}: the lexical entry, as it should appear in the head entry.
\item \code{gloss}: the gloss of the entry.
}

Optionally, the file can have the following columns:
\itemize{
\item \code{definition}: the full definition of the entry. This normally provides
more details about the meaning than the gloss. If this column is not
present, the definition field is filled with the gloss.
\item \code{phonemic}: phonemic transcription of the entry.
\item \code{phonetic}: phonetic transcription of the entry.
\item \code{word_type}: type of word (e.g. root or affix).
\item \code{word_class}: word class/part of speech of entry.
\item \code{etymology}: the etymology of the entry.
\item \code{notes}: free text notes.
}
}
