% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_lexicon_lift}
\alias{import_lexicon_lift}
\title{Import lexicon from LIFT file}
\usage{
import_lexicon_lift(lexadb, path)
}
\arguments{
\item{lexadb}{A \code{lexadb} object as returned by \code{load_lexadb()}.}

\item{path}{The path to the lexicon LIFT file as a string.}
}
\value{
Nothing. Used for its side effects.
}
\description{
It imports the lexicon from a LIFT file (exported from LexiquePro).
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
