% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{import_lexicon_mdf}
\alias{import_lexicon_mdf}
\title{Import lexicon to existing database}
\usage{
import_lexicon_mdf(lexadb, path, skip = 0)
}
\arguments{
\item{lexadb}{A \code{lexadb} object as returned by \code{load_lexadb()}.}

\item{path}{The path to the lexicon MDF file as a string.}

\item{skip}{Lines in the file to skip.}
}
\value{
A \code{lexadb} object.
}
\description{
It imports entries from a lexicon in the MDF format.
}
