% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexa-package.R
\docType{package}
\name{lexa-package}
\alias{lexa}
\alias{lexa-package}
\title{lexa: Lexicon and texts database management}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

It provides a database schema and functions to manage a language database made of a lexicon and a collection of texts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stefanocoretta/lexa}
  \item \url{https://stefanocoretta.github.io/lexa/}
  \item Report bugs at \url{https://github.com/stefanocoretta/lexa/issues}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID}) (stefanocoretta)

}
\keyword{internal}
