% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{lexa_html}
\alias{lexa_html}
\title{Lexa HTML document}
\usage{
lexa_html(...)
}
\arguments{
\item{...}{Arguments passed to \link[rmarkdown]{html_document}.}
}
\description{
A custom Rmarkdown template to typeset HTML documents with extra support for interlinear glosses. # nolint
}
