% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{lexa_pdf}
\alias{lexa_pdf}
\title{Lexa PDF document}
\usage{
lexa_pdf(...)
}
\arguments{
\item{...}{Arguments passed to \link[rmarkdown]{pdf_document}.}
}
\description{
A custom Rmarkdown template to typeset PDF documents with extra support for
interlinear glosses.
Note that this format uses a custom Pandoc LaTeX template, because of
incompatibilities bewteen expex and unicode-math.
}
