% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{lexadb_to_tbl}
\alias{lexadb_to_tbl}
\title{Convert lexicon to a tibble}
\usage{
lexadb_to_tbl(lexadb)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}
}
\value{
A tibble.
}
\description{
It converts the lexicon from a Lexa database to a tibble (data frame). Each
lexical entry is one row in the resulting tibble.
}
