% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexaR-package.R
\docType{package}
\name{lexaR-package}
\alias{lexaR}
\alias{lexaR-package}
\title{lexaR: Lexicon and texts database management}
\description{
It provides a database schema and functions to manage a language database made of a lexicon and a collection of texts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stefanocoretta/lexa}
  \item \url{https://stefanocoretta.github.io/lexa/}
  \item Report bugs at \url{https://github.com/stefanocoretta/lexa/issues}
}

}
\author{
\strong{Maintainer}: Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID}) (github: stefanocoretta)

Authors:
\itemize{
  \item Stefano Coretta \email{stefano.coretta@gmail.com} (\href{https://orcid.org/0000-0001-9627-5532}{ORCID}) (github: stefanocoretta)
}

}
\keyword{internal}
