% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_lexadb}
\alias{load_lexadb}
\title{Load a Lexa database}
\usage{
load_lexadb(path)
}
\arguments{
\item{path}{Path to Lexa database.}
}
\value{
A \code{lexadb} object.
}
\description{
Load a Lexa database
}
