% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{open_entry}
\alias{open_entry}
\title{Open a lexical entry}
\usage{
open_entry(lexadb, entry_id)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{entry_id}{A string with the entry id (the \code{lx_} prefix and leading
zeros can be omitted.)}
}
\value{
Nothing. Used for its side effects
}
\description{
It opens the file of the specified lexical entry.
}
