% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-print.R
\name{print.lexacl}
\alias{print.lexacl}
\title{Print method for collections}
\usage{
\method{print}{lexacl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lexacl}.}

\item{...}{Arguments passed to print.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Print method for objects of class \code{lexacl}, which prints collection info.
}
