% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-print.R
\name{print.lexalxscompact}
\alias{print.lexalxscompact}
\title{Compact print method for list of entries}
\usage{
\method{print}{lexalxscompact}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lexalxscompact}.}

\item{...}{Arguments passed to print.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Compact print method for the output of \code{search_lexicon()}, which returns an object
of class \code{lexalxscompact} when \code{show_entry} is \code{FALSE}..
}
