% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.lexatx}
\alias{print.lexatx}
\title{Print method for texts}
\usage{
\method{print}{lexatx}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lexatx}.}

\item{...}{Arguments passed to print.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Print method for objects of class \code{lexatx}, which prints text info.
}
