% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentences.R
\name{search_collections}
\alias{search_collections}
\title{Search words in sentence collections}
\usage{
search_collections(lexadb, word = NULL, whole = TRUE, gloss = NULL)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{word}{A regular expression to search among the sentences.}

\item{whole}{Whether to search for whole words (\code{TRUE} by default).}

\item{gloss}{A regular expression to search among the glosses.}
}
\value{
A list of \code{lexalx} objects.
}
\description{
Search words in the sentence collections.
}
\examples{
db_path <- system.file("extdata/albanian_lexadb", package = "lexa")
albanian <- load_lexadb(db_path)

search_collections(albanian, "rrezet")
search_collections(albanian, gloss = "sun")
search_collections(albanian, gloss = "traveller")
}
