% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_lexicon}
\alias{search_lexicon}
\title{Search lexicon entries}
\usage{
search_lexicon(
  lexadb,
  lexeme = NULL,
  whole = FALSE,
  definition = NULL,
  word_class = NULL,
  show_entry = FALSE
)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{lexeme}{A regular expression to search among entries.}

\item{whole}{Whether to search for whole words (only applies to \code{entry},
\code{FALSE} by default).}

\item{definition}{A regular expression to search among sense definitions.}

\item{word_class}{A regular expression to match the word class/part of speech.}

\item{show_entry}{Whether to print all the entry info (uses \code{print.lexalx}, default is `FALSE``).}
}
\value{
A list of \code{lexalx} objects.
}
\description{
Search entries in the lexicon, by entry form or sense definitions.
}
\examples{
db_path <- system.file("extdata/eleryon_lexadb", package = "lexaR")
eleryon <- load_lexadb(db_path)

# Search for "chǭs"
search_lexicon(eleryon, "chǭs")

# Search for all verbs
search_lexicon(eleryon, ".*", word_class = "verb")

# Search for entry with meaning "love"
search_lexicon(eleryon, definition = "love")
}
