% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texts.R
\name{search_texts}
\alias{search_texts}
\title{Search words in texts}
\usage{
search_texts(lexadb, word = NULL, whole = TRUE, gloss = NULL)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{word}{A regular expression to search among the sentences.}

\item{whole}{Whether to search for whole words (\code{TRUE} by default).}

\item{gloss}{A regular expression to search among the glosses.}
}
\value{
A list of \code{lexalx} objects.
}
\description{
Search words in the texts collection.
}
\examples{
db_path <- system.file("extdata/albanian_lexadb", package = "lexa")
albanian <- load_lexadb(db_path)

search_texts(albanian, "rrezet")
search_texts(albanian, gloss = "sun")
search_texts(albanian, gloss = "traveller")
}
