% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texts.R
\name{show_collection}
\alias{show_collection}
\title{Show collection or sentence with given id}
\usage{
show_collection(lexadb, coll_id, sent_id = NULL)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{coll_id}{A string with the collection id (the \code{cl_} prefix and leading
zeros can be omitted.)}

\item{sent_id}{A string with the sentence id (the \code{st_} prefix and leading
zeros can be omitted.)}
}
\value{
A \code{lexast} object.
}
\description{
It shows the collection or sentence with the given id.
}
\examples{
db_path <- system.file("extdata/albanian_lexadb", package = "lexa")
albanian <- load_lexadb(db_path)

show_collection(albanian, 1)
show_collection(albanian, 1, 3)
}
