% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texts.R
\name{show_text}
\alias{show_text}
\title{Show text or sentence with given id}
\usage{
show_text(lexadb, text_id, sent_id = NULL)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{text_id}{A string with the text id (the \code{tx_} prefix and leading
zeros can be omitted.)}

\item{sent_id}{A string with the sentence id (the \code{st_} prefix and leading
zeros can be omitted.)}
}
\value{
A \code{lexast} object.
}
\description{
It shows the text or sentence with the given id.
}
