% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloss.R
\name{typeset_gloss}
\alias{typeset_gloss}
\title{Typeset interlinear gloss}
\usage{
typeset_gloss(lexadb, text, sentence, format = "latex")
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{text}{Text id.}

\item{sentence}{Sentence id to print.}

\item{format}{Format to print out with (either \code{html} or \code{latex}).}
}
\description{
The function returns the selected sentence as properly formatted html/latex code.
}
