% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{write_merged_lexicon}
\alias{write_merged_lexicon}
\title{Write merged lexicon to disk}
\usage{
write_merged_lexicon(lexadb, sort = TRUE, overwrite = FALSE)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{sort}{Whether to sort the entries alphabetically (default is \code{TRUE}).}

\item{overwrite}{Whether to overwrite an existing \code{lexicon.yaml} file.}
}
\description{
It writes a yaml file which contains all the entries in the lexicon. The file
is output in the \code{src} directory.
}
