% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{add_entry}
\alias{add_entry}
\title{Add entry to lexicon}
\usage{
add_entry(
  lexadb,
  headword,
  gloss,
  morph_type = NULL,
  word_class = NULL,
  phonemic = NULL,
  phonetic = NULL,
  definition = gloss,
  homophone = NULL
)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{headword}{The citation form/headword of the lexical entry as a string.}

\item{gloss}{The gloss as a string.}

\item{morph_type}{The morphological type of the lexical entry (root, stem, affix, clitic, particle, compound, derived, multiword expression, phrase, ...).}

\item{word_class}{The word class/part of speech of the lexical entry.}

\item{phonemic}{The phonemic transcription.}

\item{phonetic}{The phonetic transcription.}

\item{definition}{The definition of the entry as a string.}

\item{homophone}{The homophone numeric index.}
}
\value{
Nothing. Used for its side effects
}
\description{
This function creates a new entry in the lexicon, i.e. a new empty entry
skeleton is written to disk, in the \verb{lexicon/} directory, for the user
to edit at will.
}
