% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{add_entry}
\alias{add_entry}
\title{Add entry to lexicon}
\usage{
add_entry(lexadb)
}
\arguments{
\item{lexadb}{A `lexadb` object (created with \code{\link{load_lexadb}})}
}
\value{
Nothing. Used for its side effects
}
\description{
This function creates a new entry in the lexicon, i.e. a new empty entry
skeleton is written to disk, in the `lexicon/` directory, for the user
to edit at will.
}
