% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_lexicon_csv}
\alias{import_lexicon_csv}
\title{Import lexicon from a csv file}
\usage{
import_lexicon_csv(lexadb, path)
}
\arguments{
\item{lexadb}{A \code{lexadb} object as returned by \code{load_lexadb()}.}

\item{path}{The path to the lexicon .csv file as a string.}
}
\value{
Nothing. Used for its side effects.
}
\description{
It imports entries from a \code{.csv} file with lexical data. The file must have
specific columns, see Details for file specifics. The lexicon is imported
into an existing Lexa database.
}
\details{
The file should have the following columns (they can be omitted, but
if present they must be so named):
\itemize{
\item \code{entry}: the lexical entry, as it should appear in the head entry.
\item \code{phon}: phonetic transcription of the entry.
\item \code{morph_category}: category of entry (e.g. lexical vs grammatical).
\item \code{morph_type}: type of morpheme (e.g. root vs affix).
\item \code{part_of_speech}: part of speech of entry.
\item \code{class}: lexical class of entry (e.g. verbal conjugations, noun classes).
\item \code{gloss}: .
\item \code{etymology}:
\item \code{notes}:
}

Note that this list is temporary and \emph{it will change} in the future.
}
