% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{load_lexadb}
\alias{load_lexadb}
\title{Load Lexa database}
\usage{
load_lexadb(path)
}
\arguments{
\item{path}{The path to the database as a string.}
}
\value{
A `lexadb` object.
}
\description{
It loadds a Lexa database from the specified path. The path is the directory
containing the database.
}
