% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-print.R
\name{print.lexadb}
\alias{print.lexadb}
\title{Print method for Lexa databases}
\usage{
\method{print}{lexadb}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lexadb}.}

\item{...}{Arguments passed to print.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Print method for objects of class \code{lexadb}, which prints database info and
statistics.
}
