% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.lexadb}
\alias{print.lexadb}
\title{Print method for Lexa databases}
\usage{
\method{print}{lexadb}(x, ...)
}
\arguments{
\item{x}{An object of class `lexadb`.}

\item{...}{Arguments passed to print.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Print method for objects of class `lexadb`, which prints database info and
statistics.
}
