% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\name{print.lexalx}
\alias{print.lexalx}
\title{Print method for lexemes}
\usage{
\method{print}{lexalx}(x, ...)
}
\arguments{
\item{x}{An object of class `lexalx`.}

\item{...}{Arguments passed to print.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Print method for objects of class `lexalx`, which prints lexeme info.
}
