% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-print.R
\name{print.lexast}
\alias{print.lexast}
\title{Print method for sentences}
\usage{
\method{print}{lexast}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lexast}.}

\item{...}{Arguments passed to print.}
}
\value{
Nothing. Used for its side effects.
}
\description{
Print method for objects of class \code{lexast}, which prints collection sentences.
}
