% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{search_lexicon}
\alias{search_lexicon}
\title{Search lexicon entries}
\usage{
search_lexicon(
  lexadb,
  entry = NULL,
  whole = TRUE,
  definition = NULL,
  pos = NULL
)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{entry}{A regular expression to search among entries.}

\item{whole}{Whether to search for whole words (only applies to \code{entry},
\code{TRUE} by default).}

\item{definition}{A regular expression to search among sense definitions.}

\item{pos}{A regular expression to match the part of speech.}
}
\value{
A list of \code{lexalx} objects.
}
\description{
Search entries in the lexicon, by entry form or sense definitions.
}
