% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{search_lexicon}
\alias{search_lexicon}
\title{Search lexicon entries}
\usage{
search_lexicon(lexadb, entry = NULL, definition = NULL, pos = NULL)
}
\arguments{
\item{lexadb}{A \code{lexadb} object (created with \code{\link{load_lexadb}}).}

\item{entry}{A regular expression to search across entries.}

\item{definition}{A regular expression to search across sense definitions.}

\item{pos}{A regular expression to match the part of speech.}
}
\value{
A list of \code{lexalx} objects.
}
\description{
Search entries in the lexicon, by entry form or sense definitions.
}
